chore(routes): enforce protected product workspace under /app and remove old invoice routes\n\n- Middleware: protect /app and redirect logged-in users from auth pages to /app/invoices\n- Remove legacy /dashboard/invoices and (pages)/invoices\n- Keep /app/(layout) guarded via SessionGuard
