--- src/app/dashboard/invoices/[id]/review/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function InvoiceReviewRedirect() {
-  redirect("/invoices");
-}
