--- src/app/dashboard/invoices/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function InvoicesPageRedirect() {
-  redirect("/invoices");
-}
